Added the ability to add friends, remove friends, search users to make friend requests as well as accept or decline those requests. The current loaded user is hardcoded in friends.js right now, this will change when we can read the current logged in user from browser.
PLEASE NOTE: the page for friends is static and statically served, but all friend information is generated
dynamically based on the user. This means we simply need to redirect in mainServer.js if the user is not
logged in, if they are, then we sent them to friends. Once we have the ability to log in we need to make an
endpoint to check if they are logged in so we can use that functionality when doing route logic.

I think this makes implementation the best, is follows standard practices and there aren't any security concerns.
Let me know what you think.