Feature/group chats (#22)
* Added navigation buttons. Added panel styling. Added back buttons. Added logout functionality.

* Add/remove chat implemented. Web socket iframe dynamical generation also implemented.

* Functionality that is complete:
add a new chat
delete the chat
message in the chat
load previous messages/save them as they are sent
navigate between pages
log out

Things to note:
if user A and B are in a chat and user B deletes the chat, this immediately changes on their screen, but user A would need to refresh. If user A sends more messages before they refresh though, this does nothing so functionally its okay (but adding cyclic checking would be a huge pain right now and I, myself, will not be doing it)

i did add color coding between usernames in the chat for both new and loaded messages (i.e. user A is orange and user B is red). I say this because if we want to change the coloring or styling of the text or the bubbles, not a problem, but I also will not be doing that right now since it works as is

i also added a routing handler for '/' and it checks for being logged in or not. if there is a valid cookie present in the browser, it sends the user to home, otherwise login. this only works for the index route though, we would need to do the route checking for our other pages. this was just to make it easier for booting the app since you can now ctrl+click the link in vscode and you are routed appropriately

when messages overflow in the direct messages iframe (meaning there are more messages than can fit in the window), the window will expand into a scroll window, but auto scrolling to the bottom when there are new messages doesn't work right now. there is some code in there to try and get it to work but i couldnt get it and chat didn't know what it was doing. again, not gonna worry about it because its small potatoes

* Started adding functionality based on direct messages for group chats in direct message.
Will need to add new endpoint to accept n-number of usernames or a generic "add user to chat"

* Revamped everything. Direct messages works on the same subsystem as group chats.
Group chat also have permission fields (default is 1).

* Updated mainServer to redirect to /login if no valid cookies are found with auth data.
Also handled long message wrapping in messages. [Both changes brought over from feature/check-login-redirect]

---------

Co-authored-by: eweggel74 <erw74@drexel.edu>